requirements.txt
streamlit==1.31.1
langchain==0.1.9
python-dotenv==0.21.0
google-cloud-aiplatform==1.43.0
google-generativeai==0.4.0
langchain-google-vertexai==0.1.0
unstructured==0.12.6
faiss-cpu==1.8.0